divided fxml into module
